requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
dj-database-url==2.0.0
Django==4.2.7
django-crispy-forms==2.1
djangorestframework==3.14.0
dotenv==0.9.9
gunicorn==21.2.0
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.4.4
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.5.0
